Make it possible to register modules outside of OSGi bundles
