+++ modules-registry/pom.xml
@@ -11,7 +11,7 @@
   </parent>
 
   <artifactId>modules-registry</artifactId>
-  <version>2.0.3-SNAPSHOT</version>
+  <version>2.1.0-SNAPSHOT</version>
   <packaging>bundle</packaging>
 
   <name>DAISY Pipeline 2 :: Module Registry</name>
