+++ modules-registry/src/main/java/org/daisy/pipeline/modules/impl/tracker/OSGIModuleBuilder.java
@@ -1,70 +1,16 @@
 package org.daisy.pipeline.modules.impl.tracker;
 
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-
-import org.daisy.pipeline.modules.Component;
-import org.daisy.pipeline.modules.Entity;
-import org.daisy.pipeline.modules.Module;
-import org.daisy.pipeline.modules.ResourceLoader;
-import org.daisy.pipeline.xmlcatalog.XmlCatalog;
-import org.osgi.framework.Bundle;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import com.google.common.collect.Iterators;
 
-public class OSGIModuleBuilder {
+import org.daisy.pipeline.modules.AbstractModuleBuilder;
+import org.daisy.pipeline.modules.ResourceLoader;
 
-        private ResourceLoader loader;
-        private String name;
-        private String version;
-        private String title;
-        private final List<Component> components = new ArrayList<Component>();
-        private final List<Entity> entities = new ArrayList<Entity>();
-        private final Logger mLogger = LoggerFactory.getLogger(getClass());
+import org.osgi.framework.Bundle;
 
-        public Module build() {
-                return new Module(name, version, title, components, entities);
-        }
-
-        public OSGIModuleBuilder withName(String name) {
-                this.name = name;
-                return this;
-        }
-
-        public OSGIModuleBuilder withLoader(ResourceLoader loader) {
-                this.loader = loader;
-                return this;
-        }
-
-        public OSGIModuleBuilder withVersion(String version) {
-                this.version = version;
-                return this;
-        }
-
-        public OSGIModuleBuilder withTitle(String title) {
-                this.title = title;
-                return this;
-        }
-
-        public OSGIModuleBuilder withComponents(
-                        Collection<? extends Component> components) {
-                this.components.addAll(components);
-                return this;
-        }
-
-        public OSGIModuleBuilder withComponent(URI uri, String path) {
-                mLogger.trace("withComponent:" + uri.toString() + ", path: " + path);
-                components.add(new Component(uri, path, loader));
-                return this;
-        }
+public class OSGIModuleBuilder extends AbstractModuleBuilder<OSGIModuleBuilder> {
 
         public OSGIModuleBuilder withBundle(final Bundle bundle) {
                 String title = bundle.getSymbolicName();
@@ -98,47 +44,8 @@ public class OSGIModuleBuilder {
                 return this;
         }
 
-        public OSGIModuleBuilder withCatalog(XmlCatalog catalog) {
-                for (Map.Entry<URI, URI> entry : catalog.getSystemIdMappings()
-                                .entrySet()) {
-                        withComponent(entry.getKey(), entry.getValue().toString());
-                }
-                for (Map.Entry<URI, URI> entry : catalog.getUriMappings().entrySet()) {
-                        withComponent(entry.getKey(), entry.getValue().toString());
-                }
-                for (Map.Entry<String, URI> entry : catalog.getPublicMappings()
-                                .entrySet()) {
-                        withEntity(entry.getKey(), entry.getValue().toString());
-                }
-                for (Map.Entry<URI, URI> rule : catalog.getRewriteUris().entrySet()) {
-                        Iterable<URL> entries = this.loader.loadResources(rule.getValue()
-                                        .toString());
-                        for (URL url : entries) {
-                                try {
-                                        //get tail of the path i.e. ../static/css/ -> /css/
-                                        String path=url.toURI().getPath().toString().replace(rule.getValue().toString().replace("..",""),"");
-                                        withComponent(rule.getKey().resolve(URI.create(path)), url.toString());
-                                        
-                                } catch (URISyntaxException e) {
-                                        mLogger.warn("Exception while generating paths");
-                                }
-                        }
-                }
-                
-
+        @Override
+        protected OSGIModuleBuilder self() {
                 return this;
         }
-
-        public OSGIModuleBuilder withEntities(Collection<? extends Entity> entities) {
-                this.entities.addAll(entities);
-                return this;
-        }
-        
-        public OSGIModuleBuilder withEntity(String publicId, String path) {
-                mLogger.trace("withEntity:" + publicId.toString() + ", path: " + path);
-                entities.add(new Entity(publicId, path, loader));
-                return this;
-        }
-
-
 }
